Remove thin wrapper function
